requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Cerberus==1.3.4
dj-database-url==0.5.0
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
